+++ R/get_weights.R
@@ -4,6 +4,7 @@
 #' @description Returns weighting variable of a model.
 #'
 #' @param x A fitted model.
+#' @param na_rm Logical, if \code{TRUE}, removes possible missing values.
 #' @param ... Currently not used.
 #'
 #' @return The weighting variable, or \code{NULL} if no weights were specified (or if all weights were 1).
@@ -13,14 +14,15 @@
 #' mtcars$weight <- rnorm(nrow(mtcars), 1, .3)
 #' m <- lm(mpg ~ wt + cyl + vs, data = mtcars, weights = weight)
 #' get_weights(m)
+#' @importFrom stats na.omit
 #' @export
-get_weights <- function(x, ...) {
+get_weights <- function(x, na_rm = FALSE, ...) {
   UseMethod("get_weights")
 }
 
 
 #' @export
-get_weights.default <- function(x, ...) {
+get_weights.default <- function(x, na_rm = FALSE, ...) {
   w <- NULL
   tryCatch(
     {
@@ -76,12 +78,16 @@ get_weights.default <- function(x, ...) {
     w <- NULL
   }
 
+  if (!is.null(w) && anyNA(w) && isTRUE(na_rm)) {
+    w <- stats::na.omit(w)
+  }
+
   w
 }
 
 
 #' @export
-get_weights.brmsfit <- function(x, ...) {
+get_weights.brmsfit <- function(x, na_rm = FALSE, ...) {
   w <- find_weights(x)
 
   if (!is.null(w)) {
@@ -90,6 +96,10 @@ get_weights.brmsfit <- function(x, ...) {
 
   if (!is.null(w) && all(w == 1L)) {
     w <- NULL
+  }
+
+  if (!is.null(w) && anyNA(w) && isTRUE(na_rm)) {
+    w <- stats::na.omit(w)
   }
 
   w